requirements.txt
appdirs==1.4.3
Django==1.10.6
djangorestframework==3.6.2
olefile==0.44
packaging==16.8
Pillow==4.0.0
pkg-resources==0.0.0
Pygments==2.2.0
pyparsing==2.2.0
six==1.10.0
whitenoise==3.3.0
